feat: updated folder structure and organized files
