+++ src/Pages/misc/ComingSoonPage.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
-import Footer from "../Components/Footer";
-import Header from "../Components/Header";
+import Footer from "../../components/layout/Footer.jsx";
+import Header from "../../components/layout/Header.jsx";
 
 const ComingSoonPage = () => {
   return (
