+++ src/components/common/ScrollToTop.jsx
@@ -1,4 +1,4 @@
-// src/Components/ScrollToTop.jsx
+// src/components/ScrollToTop.jsx
 
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
