+++ src/components/show/ShowCard.jsx
