+++ src/components/video/VideoPlayerUpNext.jsx
@@ -1,4 +1,4 @@
-// src/Components/VideoPlayerUpNext.jsx
+// src/components/VideoPlayerUpNext.jsx
 import { HugeiconsIcon } from "@hugeicons/react";
 import { ArrowLeft01Icon, ArrowRight01Icon } from "@hugeicons/core-free-icons";
 import { useRef } from "react";
